Struct wdc65816::Cpu

pub struct Cpu<M: Mem> {
    pub a: u16,
    pub x: u16,
    pub y: u16,
    pub s: u16,
    pub dbr: u8,
    pub pbr: u8,
    pub d: u16,
    pub pc: u16,
    pub trace: bool,
    pub mem: M,
    // some fields omitted
}

Fields
a: u16 x: u16 y: u16 s: u16

Stack pointer

dbr: u8

Data bank register. Bank for all memory accesses.

pbr: u8

Program bank register. Opcodes are fetched from this bank.

d: u16

Direct (page) register. Address offset for all instruction using "direct addressing" mode.

pc: u16

Program counter. Note that PBR is not changed on pc overflow, so code can not span multiple banks (without jml or jsr).

trace: bool mem: M
Methods

impl<M: Mem> Cpu<M>

pub fn new(mem: M) -> Cpu<M>

Creates a new CPU and executes a reset. This will fetch the RESET vector from memory and put the CPU in emulation mode.

pub fn dispatch(&mut self) -> u16

Executes a single opcode and returns the number of CPU clock cycles used.

Note that in case a WAI instruction was executed, this will not execute anything and return 0. An interrupt has to be caused to resume work.

pub fn trigger_nmi(&mut self)

Invokes the NMI handler.

pub fn trigger_irq(&mut self) -> bool

Invokes the IRQ handler if interrupts are enabled. Returns whether the interrupt was generated.